feat: Add Rust library target for mobile platform support
Tauri mobile apps (iOS/Android) require the Rust code to be compiled
as a library, not just a binary. This adds the necessary configuration
for mobile builds while keeping desktop builds working.

Changes to Cargo.toml stub:
- Added [lib] section with crate-type = ["staticlib", "cdylib", "rlib"]
- Library name follows pattern: {app_name}_lib

New lib.rs file:
- Contains main application logic
- Exports run() function marked with #[cfg_attr(mobile, tauri::mobile_entry_point)]
- Shared by both desktop and mobile builds

Updated main.rs:
- Simplified to just call {app_name}_lib::run()
- Keeps desktop binary builds working
- Code is now shared via lib.rs

Updated InitCommand:
- Copies lib.rs file during initialization

This fixes the error:
"error: no library targets found in package"

Mobile builds now properly compile the Rust code as a library,
while desktop builds continue to work as before.
